Specify dtype in max_integral

+++ python/lsst/ts/ofc/ofc_data/ofc_data.py
@@ -759,9 +759,9 @@ class OFCData(BaseOFCData):
             self.rotation_offset = self.controller["rotation_offset"]
 
         if "max_integral" in self.controller:
-            self.max_integral = np.array(self.controller["max_integral"])
+            self.max_integral = np.array(self.controller["max_integral"], dtype=float)
         else:
-            self.max_integral = np.ones(50)
+            self.max_integral = np.ones(50, dtype=float)
 
         if self.controller["name"] == "PID":
             for key in ["kp", "ki", "kd", "setpoint"]:
